replace function called by __call__ method of class for calculating overlap matrix
